pushed inbound into shared, need to abstract more state machine functionality
